Rename rta to session
